requirements.txt
PyYAML>=6.0
fastapi>=0.110
plotly>=5.0
pytest>=8.0
python-multipart>=0.0.9
uvicorn>=0.29
